Move Case Light to LEDs submenu (#18290)
