+++ Marlin/src/lcd/menu/menu_configuration.cpp
@@ -269,24 +269,6 @@ void menu_advanced_settings();
 
 #endif
 
-#if ENABLED(CASE_LIGHT_MENU)
-
-  #include "../../feature/caselight.h"
-
-  #if DISABLED(CASE_LIGHT_NO_BRIGHTNESS)
-
-    void menu_case_light() {
-      START_MENU();
-      BACK_ITEM(MSG_CONFIGURATION);
-      EDIT_ITEM(percent, MSG_CASE_LIGHT_BRIGHTNESS, &case_light_brightness, 0, 255, update_case_light, true);
-      EDIT_ITEM(bool, MSG_CASE_LIGHT, (bool*)&case_light_on, update_case_light);
-      END_MENU();
-    }
-
-  #endif
-
-#endif
-
 #if ENABLED(FWRETRACT)
 
   #include "../../feature/fwretract.h"
@@ -401,18 +383,6 @@ void menu_configuration() {
     #if ENABLED(TOOLCHANGE_MIGRATION_FEATURE)
       SUBMENU(MSG_TOOL_MIGRATION, menu_toolchange_migration);
     #endif
-  #endif
-
-  //
-  // Set Case light on/off/brightness
-  //
-  #if ENABLED(CASE_LIGHT_MENU)
-    #if DISABLED(CASE_LIGHT_NO_BRIGHTNESS)
-      if (TERN1(CASE_LIGHT_USE_NEOPIXEL, PWM_PIN(CASE_LIGHT_PIN)))
-        SUBMENU(MSG_CASE_LIGHT, menu_case_light);
-      else
-    #endif
-        EDIT_ITEM(bool, MSG_CASE_LIGHT, (bool*)&case_light_on, update_case_light);
   #endif
 
   #if HAS_LCD_CONTRAST
